Move PrimitiveBinarySerializer tests to its own file
